requirements_step1.txt
# Поэтапное восстановление зависимостей
# Начинаем с минимума для Telegram бота

python-telegram-bot==20.7
python-dotenv==1.0.0
asyncpg==0.29.0
requests==2.31.0
